+++ Installation/Installation_SoftwareInstallation-Supplementary.sh
@@ -36,28 +36,76 @@ CallScript_InstallVeraCrypt() {
     echo "VeraCrypt installation script completed!"
 }
 
-Install_Discord() {
-    read -p "Would you like to install Discord from a RPM Fusion repository or a Flathub repository? (r/f/ anything else to skip) (recommended: r): "
+Install_MicrosoftVisualStudioCode() {
+    echo "Installing Microsoft Visual Studio Code..."
 
-    if [[ $REPLY =~ ^[Rr]$ ]]
+    # Importing or creating "Microsoft Visual Studio Code" repository
+    if [[ -f "$WORKINGDIRECTORY"/Repositories/vscode.repo ]]
     then
-        echo "Installing Discord..."
-
-        sudo dnf install --assumeyes discord
-    elif [[ $REPLY =~ ^[Ff]$ ]]
-    then
-        echo "Installing Discord..."
-
-        # Installing Flatpak repository
-        sudo flatpak remote-add --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo
-
-        # Installing Discord Flatpak container
-        sudo flatpak install --assumeyes flathub com.discordapp.Discord
+        sudo cp "$WORKINGDIRECTORY"/Repositories/vscode.repo /etc/yum.repos.d/
     else
-        echo "Installing Discord skipped!"
+        sudo sh -c "printf '%s\n' \
+            '[vscode]' \
+            'name=Visual Studio Code' \
+            'baseurl=https://packages.microsoft.com/yumrepos/vscode' \
+            'enabled=1' \
+            'gpgcheck=1' \
+            'gpgkey=https://packages.microsoft.com/keys/microsoft.asc' > /etc/yum.repos.d/vscode.repo"
     fi
 
-    echo "Discord installation script completed!"
+    # Installing Microsoft Visual Studio Code from official repository
+    sudo dnf install --assumeyes code
+
+    echo "Microsoft Visual Studio Code installation script completed!"
+}
+
+Install_Skype() {
+    echo "Installing Skype..."
+
+    # Importing or creating "Skype" repository
+    if [[ -f "$WORKINGDIRECTORY"/Repositories/skype-stable.repo ]]
+    then
+        sudo cp "$WORKINGDIRECTORY"/Repositories/skype-stable.repo /etc/yum.repos.d/
+    else
+        sudo sh -c "printf '%s\n' \
+            '[skype]' \
+            'name=Skype' \
+            'baseurl=https://repo.skype.com/rpm/stable/' \
+            'enabled=1' \
+            'gpgcheck=1' \
+            'gpgkey=gpgkey=https://repo.skype.com/data/SKYPE-GPG-KEY' > /etc/yum.repos.d/skype-stable.repo"
+    fi
+
+    # Installing Skype from official repository
+    sudo dnf install --assumeyes skypeforlinux
+
+    echo "Skype installation script completed!"
+}
+
+Install_TeamViewer() {
+    echo "Installing TeamViewer..."
+
+    # Importing or creating "TeamViewer" repository
+    if [[ -f "$WORKINGDIRECTORY"/Repositories/teamviewer.repo ]]
+    then
+        sudo cp "$WORKINGDIRECTORY"/Repositories/teamviewer.repo /etc/yum.repos.d/
+    else
+        sudo sh -c "printf '%s\n' \
+            '[teamviewer]' \
+            'name=TeamViewer' \
+            'baseurl=https://linux.teamviewer.com/yum/stable/main/binary-\$basearch/' \
+            'type=rpm-md' \
+            'enabled=1' \
+            'failovermethod=priority' \
+            'gpgcheck=1' \
+            'repo-gpgcheck=1' \
+            'gpgkey=https://linux.teamviewer.com/pubkey/currentkey.asc' > /etc/yum.repos.d/teamviewer.repo"
+    fi
+
+    # Installing TeamViewer from official repository
+    sudo dnf install --assumeyes teamviewer
+
+    echo "TeamViewer installation script completed!"
 }
 
 Install_Subsurface() {
@@ -67,15 +115,26 @@ Install_Subsurface() {
     then
         echo "Installing Subsurface..."
 
-        # Importing or installing Subsurface repository
+        # Importing or installing "Subsurface" repository
         if [[ -f "$WORKINGDIRECTORY"/Repositories/subsurface.repo ]]
         then
             sudo cp "$WORKINGDIRECTORY"/Repositories/subsurface.repo /etc/yum.repos.d/
         else
-            sudo dnf config-manager --add-repo https://download.opensuse.org/repositories/home:Subsurface-Divelog/Fedora_33/home:Subsurface-Divelog.repo
+            sudo sh -c "printf '%s\n' \
+                '[subsurface]' \
+                'name=Subsurface Divelog' \
+                'baseurl=https://download.opensuse.org/repositories/home:/Subsurface-Divelog/Fedora_\$releasever/' \
+                'type=rpm-md' \
+                'enabled=1' \
+                'gpgcheck=1' \
+                'gpgkey=https://download.opensuse.org/repositories/home:/Subsurface-Divelog/Fedora_\$releasever/repodata/repomd.xml.key' > /etc/yum.repos.d/subsurface.repo"
+
+            # "Subsurface" repository official installation method
+            # RELEASEVERSION=$(awk --field-separator="=" '/VERSION_ID/ {print $2}' /etc/os-release)
+            # sudo dnf config-manager --add-repo https://download.opensuse.org/repositories/home:Subsurface-Divelog/Fedora_$RELEASEVERSION/home:Subsurface-Divelog.repo
         fi
 
-        # Installing Subsurface from repository
+        # Installing Subsurface from official repository
         sudo dnf install --assumeyes subsurface
     elif [[ $REPLY =~ ^[Ff]$ ]]
     then
@@ -163,16 +222,18 @@ CallScript_InstallDaVinciResolve() {
 Menu() {
     echo
 
-    PS3="Press 1 to exit, 2 to run all options or 3-10 to select an option to run: "
+    PS3="Press 1 to exit, 2 to run all options or 3-12 to select an option to run: "
 
-    select options in "EXIT" "RUN ALL OPTIONS" "Install WinRAR" "Install VeraCrypt" "Install Discord" "Install Subsurface" "Install Spotify" "Install Stremio" "Install WebTorrent" "Install DaVinci Resolve"; do
+    select options in "EXIT" "RUN ALL OPTIONS" "Install WinRAR" "Install VeraCrypt" "Install Microsoft Visual Studio Code" "Install Skype" "Install TeamViewer" "Install Subsurface" "Install Spotify" "Install Stremio" "Install WebTorrent" "Install DaVinci Resolve"; do
         case "$options" in
             "EXIT" )
                 exit 0;;
             "RUN ALL OPTIONS" )
                 Install_WinRAR
                 CallScript_InstallVeraCrypt
-                Install_Discord
+                Install_MicrosoftVisualStudioCode
+                Install_Skype
+                Install_TeamViewer
                 Install_Subsurface
                 Install_Spotify
                 CallScript_InstallStremio
@@ -185,8 +246,14 @@ Menu() {
             "Install VeraCrypt" )
                 CallScript_InstallVeraCrypt
                 Menu;;
-            "Install Discord" )
-                Install_Discord
+            "Install Microsoft Visual Studio Code" )
+                Install_MicrosoftVisualStudioCode
+                Menu;;
+            "Install Skype" )
+                Install_Skype
+                Menu;;
+            "Install TeamViewer" )
+                Install_Teamviewer
                 Menu;;
             "Install Subsurface" )
                 Install_Subsurface