+++ Pre-installation/Pre-installation_Export-ProgramSettings.sh
@@ -6,43 +6,36 @@
 
 
 
-Export_Settings-PulseEffects() {
-    echo "Exporting PulseEffects settings..."
+BackUp_Profiles() {
+    echo "Backing up $PROFILESNAME profiles..."
 
-    if [[ -f $WORKINGDIRECTORY/../../Backup/Backup_PulseEffectsSettings.tar.zst ]]
+    PROFILESBACKUPFILENAME=Backup_${PROFILESNAME// /}Profiles.tar.zst
+
+    if [[ -f $PROFILESDIRECTORY/$PROFILESBACKUPFILENAME ]]
     then
-        read -p "Would you like to overwrite last exported PulseEffects settings? (y/ anything else to skip): "
+        read -p "Would you like to overwrite last backed up $PROFILESNAME profiles? (y/ anything else to skip): "
 
         if [[ $REPLY =~ ^[Yy]$ ]]
         then
-            tar --create --zstd --xattrs --file="$WORKINGDIRECTORY"/../../Backup/Backup_PulseEffectsSettings.tar.zst --directory $HOME/.config/PulseEffects/ .
+            tar --create --zstd --xattrs --file="$PROFILESDIRECTORY"/$PROFILESBACKUPFILENAME --directory $PROFILES .
         else
-            echo "Exporting PulseEffect settings skipped!"
+            echo "Backing up $PROFILESNAME profiles skipped!"
         fi
     else
-        tar --create --zstd --xattrs --file="$WORKINGDIRECTORY"/../../Backup/Backup_PulseEffectsSettings.tar.zst --directory $HOME/.config/PulseEffects/ .
+        tar --create --zstd --xattrs --file="$PROFILESDIRECTORY"/$PROFILESBACKUPFILENAME --directory $PROFILES .
     fi
 
-    echo "PulseEffects settings exporting script completed!"
-}
-
-CallScript_BackUp-MozillaProfiles() {
-    echo "Calling Mozilla profiles backing up script..."
-
-    chmod +x "$WORKINGDIRECTORY"/../../Pre-installation_BackUp-MozillaProfiles.sh
-    "$WORKINGDIRECTORY"/../.././Pre-installation_BackUp-MozillaProfiles.sh
-
-    echo "Mozilla profiles backing up script completed!"
+    echo "$PROFILESNAME profiles backing up script completed!"
 }
 
 Export_Settings() {
-    echo "Exporting $SETTINGNAME..."
+    echo "Exporting $SETTINGSNAME..."
 
     for SETTING in ${SETTINGS[@]}
     do
         if [[ -f $SETTING ]]
         then
-            # Export configuration file
+            # Exporting configuration file
             if [[ -f $SETTINGSDIRECTORY/$(basename $SETTING) ]]
             then
                 read -p "Would you like to overwrite last exported \"$(basename $SETTING)\"? (y/ anything else to skip): "
@@ -56,48 +49,106 @@ Export_Settings() {
             else
                 cp $SETTING $SETTINGSDIRECTORY
             fi
+
+            # Running additional tasks
+            if [[ $SETTINGSNAME = "Git settings" ]]
+            then
+                # Backing up GitHub verification
+                if $(grep --silent "signingkey" $SETTING)
+                then
+                    PROFILESNAME="GNU Privacy Guard"
+                    PROFILES=($HOME/.gnupg)
+                    BackUp_Profiles
+                fi
+            fi
         else
-            echo "Exporting $SETTINGNAME skipped because the configuration file \"$(basename $SETTING)\" does not exist!"
+            echo "Exporting $SETTINGSNAME skipped because the configuration file \"$(basename $SETTING)\" does not exist!"
         fi
     done
 
-    echo "Exporting $SETTINGNAME script completed!"
+    echo "Exporting $SETTINGSNAME script completed!"
 }
 
 
 Menu() {
     echo
 
-    PS3="Press 1 to exit, 2 to run all options or 3-6 to select an option to run: "
+    PS3="Press 1 to exit, 2 to run all options or 3-10 to select an option to run: "
 
-    select options in "EXIT" "RUN ALL OPTIONS" "Export PulseEffects settings" "Back up Mozilla profiles" "Export Git settings" "Export Subsurface settings and database"; do
+    select options in "EXIT" "RUN ALL OPTIONS" "Back up PulseEffects profiles" "Back up Mozilla Firefox profiles" "Back up Mozilla Thunderbird profiles" "Back up LibreOffice profiles" "Export Git settings" "Export Transmission settings" "Export VLC media player settings" "Export Subsurface settings and database"; do
         case "$options" in
             "EXIT" )
                 exit 0;;
             "RUN ALL OPTIONS" )
-                Export_Settings-PulseEffects
-                CallScript_BackUp-MozillaProfiles
-                SETTINGNAME="Git settings"
+                PROFILESNAME="PulseEffects"
+                PROFILES=($HOME/.config/PulseEffects)
+                BackUp_Profiles
+                PROFILESNAME="Mozilla Firefox"
+                PROFILES=($HOME/.mozilla/firefox)
+                BackUp_Profiles
+                PROFILESNAME="Mozilla Thunderbird"
+                PROFILES=($HOME/.thunderbird)
+                BackUp_Profiles
+                PROFILESNAME="LibreOffice"
+                PROFILES=($HOME/.config/libreoffice/4/user)
+                BackUp_Profiles
+                SETTINGSNAME="Git settings"
                 SETTINGS=($HOME/.gitconfig)
                 Export_Settings
-                SETTINGNAME="Subsurface settings and database"
-                SETTINGS=($HOME/.config/Subsurface/Subsurface.conf $HOME/.subsurface/$(whoami).xml)
+                SETTINGSNAME="Transmission settings"
+                SETTINGS=($HOME/.config/transmission/settings.json)
+                Export_Settings
+                SETTINGSNAME="VLC media player settings"
+                SETTINGS=($HOME/.config/vlc/vlcrc)
+                Export_Settings
+                SETTINGSNAME="Subsurface settings"
+                SETTINGS=($HOME/.config/Subsurface/Subsurface.conf)
+                Export_Settings
+                SETTINGSNAME="Subsurface database"
+                SETTINGS=($HOME/.subsurface/$(whoami).xml)
                 Export_Settings
                 exit 0;;
-            "Export PulseEffects settings" )
-                Export_Settings-PulseEffects
+            "Back up PulseEffects profiles" )
+                PROFILESNAME="PulseEffects"
+                PROFILES=($HOME/.config/PulseEffects)
+                BackUp_Profiles
                 Menu;;
-            "Back up Mozilla profiles" )
-                CallScript_BackUp-MozillaProfiles
+            "Back up Mozilla Firefox profiles" )
+                PROFILESNAME="Mozilla Firefox"
+                PROFILES=($HOME/.mozilla/firefox)
+                BackUp_Profiles
+                Menu;;
+            "Back up Mozilla Thunderbird profiles" )
+                PROFILESNAME="Mozilla Thunderbird"
+                PROFILES=($HOME/.thunderbird)
+                BackUp_Profiles
+                Menu;;
+            "Back up LibreOffice profiles" )
+                PROFILESNAME="LibreOffice"
+                PROFILES=($HOME/.config/libreoffice/4/user)
+                BackUp_Profiles
                 Menu;;
             "Export Git settings" )
-                SETTINGNAME="Git settings"
+                SETTINGSNAME="Git settings"
                 SETTINGS=($HOME/.gitconfig)
+                Export_Settings
+                Menu;;
+            "Export Transmission settings" )
+                SETTINGSNAME="Transmission settings"
+                SETTINGS=($HOME/.config/transmission/settings.json)
+                Export_Settings
+                Menu;;
+            "Export VLC media player settings" )
+                SETTINGSNAME="VLC media player settings"
+                SETTINGS=($HOME/.config/vlc/vlcrc)
                 Export_Settings
                 Menu;;
             "Export Subsurface settings and database" )
-                SETTINGNAME="Subsurface settings and database"
-                SETTINGS=($HOME/.config/Subsurface/Subsurface.conf $HOME/.subsurface/$(whoami).xml)
+                SETTINGSNAME="Subsurface settings"
+                SETTINGS=($HOME/.config/Subsurface/Subsurface.conf)
+                Export_Settings
+                SETTINGSNAME="Subsurface database"
+                SETTINGS=($HOME/.subsurface/$(whoami).xml)
                 Export_Settings
                 Menu;;
         esac
@@ -108,6 +159,7 @@ Menu() {
 
 COLUMNS=1
 WORKINGDIRECTORY="$(pwd)/Pre-installation/Settings/User"
+PROFILESDIRECTORY="$(pwd)/Pre-installation/Backup"
 SETTINGSDIRECTORY="$WORKINGDIRECTORY"
 
 Menu